Added a service module where the buisness logic in controller module is refactored to, by doing so, multiple get requests are handled by a single imported function in controller module.
