+++ src/controller/employeeController.js
@@ -1,8 +1,10 @@
 const express = require("express");
 const app = express();
+exports.app = app;
 const employeeDAO = require("../models/employeeRepository");
 const bodyParser = require("body-parser");
 const { port, host } = require("./controller-config");
+const { handleGetRequest } = require("../service/employeeService");
 
 app.use(bodyParser.json());
 
@@ -18,6 +20,7 @@ const path = {
   UPDATE: {},
   DELETE: {},
 };
+exports.path = path;
 const HTTP_STATUS_CODE = {
   OK: 200,
   INTERNAL_SERVER_ERROR: 500,
@@ -27,90 +30,25 @@ const HTTP_STATUS_CODE = {
   SERVICE_UNAVIALABLE: 503,
 };
 
-const STATUS_MESSAGE = {};
+
+const STATUS_MESSAGE = {
+  NOT_FOUND: (data) => `${data} does not exist in the database`,
+};
 
 app.get(path.GET.base, (_, response) => {
-  response.contentType("json");
-  employeeDAO
-    .getAllEmployees()
-    .then((data) => {
-      response.json(data);
-      response.statusCode = HTTP_STATUS_CODE.OK;
-    })
-    .catch((error) => {
-      response.statusCode = HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR;
-      response.statusMessage = `ERROR fetching data from database \n ${error.message}`;
-    })
-    .finally(() => {
-      response.end();
-    });
+  handleGetRequest(response, employeeDAO.getAllEmployees());
 });
 // route parameters
 app.get(path.GET.email, (request, response) => {
   const { email } = request.params;
   if (email) {
-    employeeDAO
-      .getEmployeeByEmail(email)
-      .then((data) => {
-        response.json(data);
-        response.statusCode = HTTP_STATUS_CODE.OK;
-      })
-      .catch((error) => {
-        response.statusCode = HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR;
-        response.statusMessage = `Error fetching email from database -> \n${error.message}`;
-      })
-      .finally(response.end);
-  }
-});
-
-app.post(path.POST.addEmployee, (request, response) => {
-  const {
-    name,
-    company_name: companyName,
-    role,
-    salary,
-    phone_number: phoneNumber,
-    email,
-  } = request.body;
-
-  const employee = {
-    name,
-    companyName,
-    role,
-    salary,
-    phoneNumber,
-    email,
-  };
-  try {
-    employeeDAO.addEmployee(employee);
-    response.status = HTTP_STATUS_CODE.OK;
-    response.statusMessage = "Employee added to database successfully";
-  } catch (error) {
-    response.status = HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR;
-    response.statusMessage = error.message;
-  } finally {
-    response.end();
+    handleGetRequest(response, employeeDAO.getEmployeeByEmail(email));
+  } else {
+    response.statusCode = HTTP_STATUS_CODE.NOT_FOUND;
+    response.statusMessage = STATUS_MESSAGE.NOT_FOUND("Email");
   }
 });
 
 app.listen(port, () => console.log(`Express is on `));
 
-/**
- * @param {Function(any):Promise} callback
- * @param {Express.Response} response
- * @param {*} [arg]
- * @returns {void}
- */
-function handleGetRequest(response, callback, arg) {
-  response.contentType('json')
-  callback(arg)
-    .then((result) => {
-      response.statusCode = HTTP_STATUS_CODE.OK;
-      response.json(result);
-    })
-    .catch((err) => {
-      response.statusCode = HTTP_STATUS_CODE.INTERNAL_SERVER_ERROR;
-      response.statusMessage = err.message;
-    })
-    .finally (response.end);
-}
+exports.HTTP_STATUS_CODE = HTTP_STATUS_CODE;
